requirements.txt
numpy==1.26.4
pygame-ce==2.5.6
opencv-python==4.12.0.88
mediapipe==0.10.21